Temporary fix for windows 2019 agent CI build
The go docker image with tag 1.23 no longer has windows server 2019 as one of
its platforms. This is a temporary fix to allow the agent to be built. It
still uses the current latest go 1.23 version as that is still available for
windows server 2019, but it has to be specified manually since it's no longer
part of the multi-platform image 1.23.